+++ NEWS.md
@@ -1,6 +1,6 @@
 # ggpedigree 0.5.1
 * Fix tooltip appearing in `ggPedigreeInteractive` when tooltip_include is set to FALSE.
-* Making progress on filling by matID, patID
+* Implemented fill by matID, patID using several scales, such as hue, viridis, and others.
 
 # ggpedigree 0.5.0
 * Added segment_linetype and custom affected labels to the `ggpedigree` function.